Renamed the JobPaths class with FileSystem as it is a more accurate description of the class' purpose (see issue #2) Moved the 'fetch_build' methods from execution_script and placed it in the FileSystem class Replaced all job_path references with file_system
